Capture full session on alert - Feature #120
In most signature based IDS's, when a signature fires, they have the
ability to dump the packet data for which the signature fired upon
and subsequent packet data, usually within the same network stream.
Unfortunately, this can mean important information may not be
available to a user if the information occured prior to the packet
containing the signature.  For example, if a signature fires on an
HTTP response, it is possible that important information in an
HTTP request may not be available but could be useful in
investigations.

TimeMachine is a fast, memory-efficient solution for storing network
traffic for retrieval after an alarm fires and dumping all stored
captured traffic to disk.

When TimeMachine is enabled, each detection thread creates one or more
pre-allocated in-memory heaps based on a configurable packet size and
configurable amount of memory to use.

When a new packet is processed by a given thread, several actions may
take place:

  -  In cases where TimeMachine does not know of an existing flow
     associated with a given packet, TimeMachine will generate a
     new minimal flow stored within a splay tree, store the packet
     data in one of the pre-allocated in-memory heaps and associate
     the packet to the given flow.

  -  In cases where TimeMachine already knows of an existing flow,
     it once again stores the data in the pre-allocated in-memory
     heap and associates the packet to the given flow.  This process
     also rebalances the splay tree ensuring the most recent flows
     should take less time to find within the tree structure

  -  In cases where the packet is associated with a given alert
     generated by the detection engine, a new output file will be
     opened and all data associated with the packet's flow will be
     dumped to disk in chronological order.

  -  In cases where an output file already exists indicating the
     flow was previously associated with an alert, packets are not
     stored within the in-memory heap and instead are written
     directly to disk.

As the heaps are fixed size, when a given heap is full, TimeMachine
will purge the oldest packet first.  In cases where the packet is
the last item associated with a given flow, TimeMachine will also
clean up the given flow information.

TimeMachine is an effective alternative in cases where full packet
capture is not available due to disk or IOP's constraints.  The
overall effectiveness of TimeMachine is dependent upon the overall
network speed, size of heaps configured for TimeMachine and total
memory available.

This is an implementation to address feature request #120, capture
full session on alert:

https://redmine.openinfosecfoundation.org/issues/120