stream: fix logic problem in midstream
When midstream is enabled, picked up flow are sent to app-layer
that has to recognize the flow.
The code is analyzing the data in to_server direction
and continue in to_client using the same approach.
In this case the port are not inverted although
we consider inverted data.
The result is that pe1 and pe2 are NULL due to the inversion.

To fix that error, the patch inverts source and destination ports
to pick up the connection when we are in to_client direction.

Consequences of the error was that all stream starting
with a to_client message where not picked up.

In the case of the modbus protocol where almost every packet is one command,
suricata was failing to pickup the connections as soon as it was starting
with a to_client message.
Run on forged pcap files have shown that, with this modification,
suricata is behaving correctly.