stream-tcp: fix logic problem in midstream
When midstream is enabled, picked up flow are sent to app-layer
that has to recognize the flow. The code is analyzing the data
in to_server direction and continue in to_client using the same
approach.
In this case the port are not inverted although we consider
inverted data. The result is that both probers are NULL due to
the inversion.

Consequences of the error was that all stream starting with a
to_server message were not analyzed correctly by application layer.

To fix that error, the patch inverts ports when source port
is tested and mark the flow has reverted. This supposes only
destination port are registered as probing port.

When flow is reverted and midstream is enable, we run the
opposite application layer parser to assume the consequence of
our previous choice.

In the case of the modbus protocol where almost every packet is one
command, suricata was failing to pickup the connections as soon as
it was starting with a to_server message.

Run on forged pcap files have shown that, with this modification,
suricata is behaving correctly.