Issue 1592: Fix deadlock in thread restart
This fixes an issues where when a thread dies and attempts to be
restarted, a deadlock ensues as the thread attempts to lock the same
mutex (tv_root_lock) twice.

Specifically, TmThreadCheckThreadState() takes a lock on tv_root_lock as
it iterates through the list of threads and performs a check as to the
state of the thread.  If the thread has failed, it will attempt to
potentially restart the thread calling TmThreadRestartThread() which
subsequently calls TmThreadSpawn() which eventually calls
TmThreadAppend().  TmThreadAppend() also attempts to take a lock on
tv_root_mutex.  Since the lock was already taken by
TmThreadCheckThreadState(), a deadlock ensues.

The potential fix is to change tv_root_mutex from a normal mutex into
a recursive mutex so the same thread can lock tv_root_mutex multiple
times (this looked safe when testing).  However, there are a number
of functions that are either called once or never called that interact
with tv_root_mutex as well (example TmThreadAppend() is only called
from TmThreadSpawn() and TmThreadRestart() isn't called at all).
In both of these cases, since the function interacts with
tv_root_mutex, they create a lock.  We may want to revisit this in
general but the proposed fix should unblock the initial issue.

The change requires an explicit initialization function which is
now called in Suricata's main().