app-layer-modbus: fix transaction done detection
It we have a transaction in replied state after an unreplied
transaction, this means server did miss an answer and we can
assume we will never see it so the transaction is incomplete
but done.

THe problem was causing to delay the logging of all transactions
in a problematic flow at the end of the flow. Furthermore this was
resulting in invalid timestamping (using last packet timestamp
for all the remaining events).