bypass: bypass encrypted flow
If a flow is encrypted, suricata is stopping the inspection
so we can just get it out via bypass.

This patch also invert to test to check the option value.
This should limit the number of tests in conditional.

For a basic test of feature, use the following ruleset:

```
table ip filter {
	chain output {
		 type filter hook output priority 0;
		 meta mark set ct mark counter
		 mark 0x00000001 counter accept
		 oif lo tcp dport ssh counter queue num 0
		 oif lo tcp sport ssh counter queue num 0
	}

	chain connmark_save {
		 type filter hook output priority 1;
		 mark 0x00000001 counter
		 ct mark set mark counter
	}
}
```

And use mark and mask of 1 in nfq configuration. Then you
can test the system by scp big file to 127.0.0.1.