packet logging: improve already logged detection
Introduce a new flag field in Packet structure to contain information
from upper layer. Upper layer (like Flow) will set bit in the flag.
Interest will be to have higher level information in the Packet that
will be accessible without locking the Flow.

First usage of the flag is done in TLS and SSH JSON logging. Impact
is suppression of 2 locks and unlocks of the Flow for all TCP packets
(one to check TLS logging, and one to check SSH logging).

Second modification consist in copying alproto to packet. By using this
we avoid to relock the flow for all packets.

Perf top without optimization:
+   13.79%    13.79%  suricata                  [.] FlowManager
-   10.92%    10.92%  libpthread-2.19.so        [.] pthread_mutex_unlock
   - pthread_mutex_unlock
      + 45.39% 0
      + 6.37% JsonTlsCondition
      + 5.58% JsonSshCondition

perf top with optimization:
+   11.95%    11.95%  suricata               [.] FlowManager
-   10.40%    10.40%  libpthread-2.19.so     [.] pthread_mutex_unlock
   - pthread_mutex_unlock
      + 46.92% 0
        2.53% 0x1629d310