app-layer: add ThreadVars pointer to AppLayerParserParse
To be able to have per application layer counters, we need to get
a ThreadVars in the parsing function. The consequence is that
the AppLayerParser function has to be updated to be able to transmit
the pointer to called functions.

Modification done via the following semantic patch and manual edit:
 @@
 expression e1, e2, e3, e4, e5, e6;
 @@

 - AppLayerParserParse(e1, e2, e3, e4, e5, e6)
 + AppLayerParserParse(NULL, e1, e2, e3, e4, e5, e6)